add an activity which controlls dynamic feature module
